feat: 결제내역 모달형식으로 추가
-결제내역은 예약내역에서 모달로 띄움
-chore:폴더구조 정리
