+++ src/provider/Provider.tsx
@@ -1,0 +1,12 @@
+import QueryProvider from './QueryProvider';
+import AuthProvider from './AuthProvider';
+
+const Provider = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <QueryProvider>
+      <AuthProvider>{children}</AuthProvider>
+    </QueryProvider>
+  );
+};
+
+export default Provider;
