+++ src/stores/usePaymentStore.ts
@@ -1,0 +1,11 @@
+import { create } from 'zustand';
+
+interface PaymentState {
+  paymentModal: boolean;
+  togglePayment: () => void;
+}
+
+export const usePaymentStore = create<PaymentState>((set) => ({
+  paymentModal: false,
+  togglePayment: () => set((state) => ({ paymentModal: !state.paymentModal })),
+}));
